requirements.txt
tensorflow
numpy
pandas
opencv-python
Pillow
matplotlib
scikit-learn